TVB-2278 Code review
