+++ tvb/interfaces/web/static/style/section_connectivity.css
@@ -1,5 +1,5 @@
 /**
- * TheVirtualBrain-Framework Package. This package holds all Data Management, and 
+ * TheVirtualBrain-Framework Package. This package holds all Data Management, and
  * Web-UI helpful to run brain-simulations. To use it, you also need do download
  * TheVirtualBrain-Scientific Package (for simulators). See content of the
  * documentation-folder for more details. See also http://www.thevirtualbrain.org
@@ -662,353 +662,4 @@ margin-right: 2px;
  * Otherwise more specific styles will cascade them */
 #main .view-column .action.action-new-connectivity{
     right:40px;
-}
-
-.node {
-  font: 300 11px "Helvetica Neue", Helvetica, Arial, sans-serif;
-  fill: #bbb;
-}
-
-.node:hover {
-  fill: #000;
-}
-
-.link {
-  stroke: steelblue;
-  stroke-opacity: 0.4;
-  fill: none;
-  pointer-events: none;
-}
-
-.node:hover,
-.node--source,
-.node--target {
-  font-weight: 700;
-}
-
-.node--source {
-  fill: #2ca02c;
-}
-
-.node--target {
-  fill: #d62728;
-}
-
-.link--source,
-.link--target {
-  stroke-opacity: 1;
-  stroke-width: 2px;
-}
-
-.link--source {
-  stroke: #d62728;
-}
-
-.link--target {
-  stroke: #2ca02c;
-}
-
-.diagram-svg {
-  width: 100%;
-  height: 100%;
-  display: block;
-  margin: auto;
-}
-
-.switch-input {
-  display: none;
-}
-.switch-label {
-  position: relative;
-  display: inline-block;
-  min-width: 112px;
-  cursor: pointer;
-  font-weight: 500;
-  text-align: left;
-  margin: 16px;
-  padding: 16px 0 16px 44px;
-}
-.switch-label:before, .switch-label:after {
-  content: "";
-  position: absolute;
-  margin: 0;
-  outline: 0;
-  top: 50%;
-  -ms-transform: translate(0, -50%);
-  -webkit-transform: translate(0, -50%);
-  transform: translate(0, -50%);
-  -webkit-transition: all 0.3s ease;
-  transition: all 0.3s ease;
-}
-.switch-label:before {
-  left: 1px;
-  width: 34px;
-  height: 14px;
-  background-color: #cdcbcb;
-  border-radius: 8px;
-}
-.switch-label:after {
-  left: 0;
-  width: 20px;
-  height: 20px;
-  background-color: #7c7b7b;
-  border-radius: 50%;
-  box-shadow: 0 3px 1px -2px rgba(0, 0, 0, 0.14), 0 2px 2px 0 rgba(0, 0, 0, 0.098), 0 1px 5px 0 rgba(0, 0, 0, 0.084);
-}
-.switch-label .toggle--on {
-  display: none;
-}
-.switch-label .toggle--off {
-  display: inline-block;
-}
-.switch-input:checked + .switch-label:before {
-  background-color: #7c7b7b;
-}
-.switch-input:checked + .switch-label:after {
-  background-color: #cdcbcb;
-  -ms-transform: translate(80%, -50%);
-  -webkit-transform: translate(80%, -50%);
-  transform: translate(80%, -50%);
-}
-.switch-input:checked + .switch-label .toggle--on {
-  display: inline-block;
-}
-.switch-input:checked + .switch-label .toggle--off {
-  display: none;
-}
-
-.container {
-  width: 98%;
-  height: 99%;
-}
-
-.leftpane {
-    width: 20%;
-    height: 87%;
-    float: left;
-    background-color: #ffffff;
-    border-collapse: collapse;
-}
-
-.middlepane {
-    width: 60%;
-    height: 87%;
-    float: left;
-    background-color: #ffffff;
-    border-collapse: collapse;
-}
-
-.rightpane {
-  width: 20%;
-  height: 87%;
-  position: relative;
-  float: right;
-  background-color: #ffffff;
-  border-collapse: collapse;
-}
-
-.toppane {
-  width: 100%;
-  height: 85px;
-  border-collapse: collapse;
-  background-color: #4169e1;
-    position: relative;
-    display: table;
-
-}
-
-/*
- * slider
- */
-/* Import */
-@import url("https://fonts.googleapis.com/css?family=Sniglet|Oswald");
-@import url('https://fonts.googleapis.com/css?family=Roboto');
-
-#slider-value {
-    font-family: 'Roboto', sans-serif;
-    top: 50%; left: 45%; position: relative;
-    font-weight: bold;
-    font-size: large;
-    color: white;
-}
-
-#slider {
-  -webkit-appearance: none;
-  width: 160px;
-  height: 20px;
-  background: linear-gradient(to right, #16a085 0%, #96c869 100%);
-  background-size: 150px 10px;
-  background-position: center;
-  background-repeat: no-repeat;
-  overflow: hidden;
-  outline: none;
-  zoom: 100%;
-  display: block;
-  margin: auto;
-}
-
-#slider::-webkit-slider-thumb {
-  -webkit-appearance: none;
-  width: 20px;
-  height: 20px;
-  background: #27ae60;
-  position: relative;
-  z-index: 3;
-}
-
-#slider::-webkit-slider-thumb:after {
-  content: " ";
-  width: 160px;
-  height: 10px;
-  position: absolute;
-  z-index: 1;
-  right: 20px;
-  top: 5px;
-  background: #2ecc71;
-}
-
-.blue::-webkit-slider-thumb {
-  background: #2980b9;
-}
-
-.blue::-webkit-slider-thumb:after {
-  background: #3498db;
-}
-
-.one {
-  width: 0px;
-  height: 0px;
-  background-color: transparent;
-  line-height: 30px;
-  color: #ecf0f1;
-  cursor: pointer;
-  border: 70px solid transparent;
-  border-left: 70px solid #16a085;
-  border-top: 70px solid #16a085;
-  left: 0px;
-  top: 0px;
-  position: absolute;
-}
-
-a {
-  text-decoration: none;
-}
-
-.one span {
-  margin-left: -60px;
-  position: absolute;
-  margin-top: -40px;
-}
-
-.two {
-  width: 0px;
-  height: 0px;
-  background-color: transparent;
-  line-height: 30px;
-  color: #ecf0f1;
-  cursor: pointer;
-  border: 70px solid transparent;
-  border-right: 70px solid #16a085;
-  border-top: 70px solid #16a085;
-  right: 0px;
-  top: 0px;
-  position: absolute;
-}
-
-.two span {
-  margin-left: -15px;
-  position: absolute;
-  margin-top: -40px;
-  white-space: nowrap;
-}
-
-/*
- * panes in toppane
- */
-
-#inner-c1{
-    float:left;width:33%;
-    margin-left: auto;
-    margin-right: auto;
-}
-
-#inner-c3{
-    float:right;
-    width:auto;
-    margin-left: auto;
-    margin-right: auto;
-}
-
-#c1, #c3 {
-    display: table-cell;
-    vertical-align: middle;
-}
-
-#c2 {
-    display: table-cell;
-    vertical-align: middle;
-    margin:auto auto;
-    width:33%;
-}
-
-.select, .label{
-    display: block;
-    margin: auto;
-}
-
-.label{
-    text-align: center;
-}
-
-/*
- * select boxes
- */
-
-@import url(https://fonts.googleapis.com/css?family=Source+Sans+Pro);
-
-.select-box {
-  cursor: pointer;
-  position : relative;
-  max-width:  20em;
-  margin: 5em auto;
-  width: 100%;
-}
-
-.select,
-.label {
-  color: #414141;
-  font: 400 17px/2em 'Source Sans Pro', sans-serif;
-}
-
-.select {
-  width: 100%;
-  position: absolute;
-  top: 0;
-  padding: 5px 0;
-  height: 100%;
-  opacity: 0;
-  -ms-filter: "progid:DXImageTransform.Microsoft.Alpha(Opacity=0)";
-  background: none transparent;
-  border: 0 none;
-}
-.select-box1, .select-box2 {
-  background: #ececec;
-}
-
-.label {
-  position: relative;
-  padding: 5px 10px;
-  cursor: pointer;
-}
-.open .label::after {
-   content: "▲";
-}
-.label::after {
-  content: "▼";
-  font-size: 12px;
-  position: absolute;
-  right: 0;
-  top: 0;
-  padding: 5px 15px;
-  border-left: 5px solid #fff;
 }
