--- tvb/interfaces/web/templates/genshi/visualizers/chord/chord_view.html
@@ -1,28 +1,0 @@
-<div id="s-brain">
-    <link rel="stylesheet" type="text/css" href="/static/style/section_connectivity.css"/>
-    <script src="/static/js/d3.v4.min.js"></script>
-    <script type="text/javascript" src="/static_view/chord/chordScript.js?$currentVersionJS"></script>
-    <script type="text/javascript" src="/static_view/commons/scripts/hierarchicalEdgeBundle.js?$currentVersionJS"></script>
-    <script type="text/javascript">
-        $('document').ready(function () {
-            init_chord("${url_base}", ${labels});
-            ajaxify();
-        })
-    </script>
-    <div class="container">
-        <div class="toppane">
-            <div id="slider">
-                <input type="checkbox" id="switch-1" class="switch-input"/>
-                <label for="switch-1" class="switch-label"><span class="toggle--on">Tract Length</span><span class="toggle--off">Weights</span></label>
-            </div>
-        </div>
-        <div class="leftpane">
-
-        </div>
-        <div class="middlepane">
-            <svg viewBox="0 0 700 700" preserveAspectRatio="xMidYMid meet" class="diagram-svg" id="middle-chord"></svg>
-        </div>
-        <div class="rightpane">
-        </div>
-    </div>
-</div>
