+++ tvb/interfaces/web/templates/genshi/visualizers/connectivity_edge_bundle/view.html
@@ -1,0 +1,23 @@
+<div>
+    <link rel="stylesheet" type="text/css" href="/static/style/subsection_edge_bundle.css"/>
+    <script type="text/javascript" src="/static/js/d3.v4.min.js"></script>
+    <script type="text/javascript" src="/static_view/commons/scripts/hierarchicalEdgeBundle.js?$currentVersionJS"></script>
+    <script type="text/javascript" src="/static_view/connectivity_edge_bundle/connectivity_edge_bundle.js?$currentVersionJS"></script>
+    <script type="text/javascript">
+        $('document').ready(function () {
+            init_chord("${url_base}", ${labels});
+            $("#main").addClass("colscheme-1");
+        })
+    </script>
+    <div class="view-column">
+        <ul class="toolbar-inline">
+            <li>
+                <input type="checkbox" id="switch-1" class="switch-input"/>
+                <label for="switch-1" class="switch-label"><span class="toggle--on">Tract Length</span><span
+                        class="toggle--off">Weights</span></label>
+            </li>
+        </ul>
+
+        <svg viewBox="0 0 700 790" preserveAspectRatio="xMidYMax meet" class="diagram-svg" id="middle-edge-bundle"></svg>
+    </div>
+</div>
